break up scope keeper file
