+++ x/metadata/keeper/record.go
@@ -1,0 +1,79 @@
+package keeper
+
+import (
+	"fmt"
+
+	"github.com/provenance-io/provenance/x/metadata/types"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+// GetRecord returns the record with the given id.
+func (k Keeper) GetRecord(ctx sdk.Context, id types.MetadataAddress) (group types.Record, found bool) {
+	if !id.IsRecordAddress() {
+		return group, false
+	}
+	store := ctx.KVStore(k.storeKey)
+	b := store.Get(id)
+	if b == nil {
+		return types.Record{}, false
+	}
+	k.cdc.MustUnmarshalBinaryBare(b, &group)
+	return group, true
+}
+
+// SetRecord stores a group in the module kv store.
+func (k Keeper) SetRecord(ctx sdk.Context, record types.Record) {
+	store := ctx.KVStore(k.storeKey)
+	b := k.cdc.MustMarshalBinaryBare(&record)
+	eventType := types.EventTypeRecordCreated
+
+	recordID := record.GroupId.GetRecordAddress(record.Name)
+	if store.Has(recordID) {
+		eventType = types.EventTypeRecordUpdated
+	}
+
+	store.Set(recordID, b)
+
+	ctx.EventManager().EmitEvent(
+		sdk.NewEvent(
+			eventType,
+			sdk.NewAttribute(types.AttributeKeyScopeID, recordID.String()),
+		),
+	)
+}
+
+// RemoveRecord removes a scope from the module kv store.
+func (k Keeper) RemoveRecord(ctx sdk.Context, id types.MetadataAddress) {
+	if !id.IsRecordAddress() {
+		panic(fmt.Errorf("invalid address, address must be for a record"))
+	}
+	store := ctx.KVStore(k.storeKey)
+	store.Delete(id)
+
+	ctx.EventManager().EmitEvent(
+		sdk.NewEvent(
+			types.EventTypeRecordRemoved,
+			sdk.NewAttribute(types.AttributeKeyScopeID, id.String()),
+		),
+	)
+}
+
+// IterateRecords processes all records in a scope with the given handler.
+func (k Keeper) IterateRecords(ctx sdk.Context, scopeID types.MetadataAddress, handler func(types.Record) (stop bool)) error {
+	store := ctx.KVStore(k.storeKey)
+	prefix, err := scopeID.ScopeGroupIteratorPrefix()
+	if err != nil {
+		return err
+	}
+	it := sdk.KVStorePrefixIterator(store, prefix)
+	defer it.Close()
+	for ; it.Valid(); it.Next() {
+		var record types.Record
+		k.cdc.MustUnmarshalBinaryBare(it.Value(), &record)
+		if handler(record) {
+			break
+		}
+	}
+	return nil
+}
